% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.comparison.R
\name{by.comparison}
\alias{by.comparison}
\title{Reorder data.frame for use in nma.prior}
\usage{
\method{by}{comparison}(data.nma, outcome, type.outcome = "binomial", N, sd = NULL, time = NULL)
}
\arguments{
\item{data.nma}{A data object}

\item{outcome}{String indicating column name of outcome variable}

\item{type.outcome}{String indicating the type of outcome}

\item{N}{String indicating column name of sample size}

\item{sd}{A string for the column name containing the standard errors (only required for continuous outcomes)}

\item{time}{A string for the column names containing the follow-up times or person-time at risk for rate or rate2 data}
}
\value{
The rearranged data
}
\description{
This function was originally an internal function to BUGSnet (https://github.com/audrey-b/BUGSnet). This version was altered to avoid warnings during the CRAN check but is otherwise exactly as presented in BUGSnet.
}
