% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.viz.r
\name{comparison.viz}
\alias{comparison.viz}
\title{Visualize pairwise treatment comparisons}
\usage{
comparison.viz(data, cutoff, metric)
}
\arguments{
\item{data}{The output of the \code{net.tab()}function applied to a dataset of interest.}

\item{cutoff}{The number of pairwise treatments to be plotted. Default is all.}

\item{metric}{Metric used to determine which pairwise treatments to be included based on cutoff. The default 'cutoff' number of
treatments with the highest number of patients. Any numeric column name of the \code{net.tab()} object may be used, specified as a string.}
}
\value{
One or two ggplot objects.
}
\description{
Plots two graphs from the output of a \code{net.tab()} function call.
In the first graph, the number of total patients and observed events are plotted for all treatment pairs in the network.
This data is pooled from all studies, and organized by each pairwise comparison among treatments of interest.
This graph is omitted for continous response data since the number of events is not meaningful for this response type.
In the second graph, the number of studies which examined each treatment pair (summed across the network) is plotted as a bar graph.
The number of comparisons made can be controlled with the cutoff and metric parameters, as in \code{intervention.viz()}
Default value for these arguments are all pairwise treatments, and the number of events (descending), respectively.
}
\examples{
\dontrun{
data("diabetes.sim")
data_diabetes <- data.prep(arm.data = diabetes.sim, varname.t = "Treatment", varname.s = "Study")
summary_diabetes <- net.tab(data = data_diabetes,
                          outcome = "diabetes", N = "n", type.outcome = "rate2", time = "followup")
comparison.viz(summary_diabetes)
}

}
