% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fenormlogpost.R
\name{fenormlogpost}
\alias{fenormlogpost}
\title{Calculate log-posterior in R}
\usage{
fenormlogpost(testPars, stanObj)
}
\arguments{
\item{testPars}{A named list of parameters to use}

\item{stanObj}{A StanNetRun object to supply the data}
}
\value{
The value of the log-posterior distribution at the test parameters and data
}
\description{
Computes the log-posterior for a fixed effect model with normal likelihood using the identity link and default priors
}
