% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_viz.r
\name{intervention_viz}
\alias{intervention_viz}
\title{Visualize individual treatment results.}
\usage{
intervention_viz(data, cutoff, type.outcome, metric)
}
\arguments{
\item{data}{The output of the \code{net.tab()} function applied to a dataset of interest. Must be specified.}

\item{cutoff}{The number of pairwise treatments to be plotted. Default is all.
If a different value is specified, the cutoff in combination with the specified metric is used to select treatments.}

\item{type.outcome}{A string which denotes the type of response variable examined in each study, similar to the net.tab specification.
Options are 'binary' (binary data with no follow-up), 'rate' (count data with person-time-at-risk), 'rate2' (binary data with follow-up time) and 'continuous'}

\item{metric}{A string which specifies the metric used to determine which pairwise treatments to be included based on cutoff.
The default metric used is the maximum number of patients.
Any numeric column name of the net.tab object may be used for this parameter.}
}
\value{
Two or three ggplot objects.
}
\description{
Plots three graphs using data from a \code{net.tab()} object.
In the first graph, the average outcome or average events per person (depending on type of response data) is plotted against treatment.
In the second graph, the number of studies which examined each treatment is plotted.
In the third graph, the number of events relative to the total number of patients for each treatment across all studies.
Note that the third graph does not appear for continuous response data.
The number of comparisons made can be controlled with the cutoff and metric parameters.
Default value for these arguments are all ptreatments, and the number of patients (descending), respectively.
The type.outcome parameter must be specified, which is the same as given in the \code{net.tab} documentation.
}
\examples{
\dontrun{
data("diabetes.sim")
data_diabetes <- data.prep(arm.data = diabetes.sim,
varname.t = "Treatment",
varname.s = "Study")
summary_diabetes <- net.tab(data = data_diabetes,
                          outcome = "diabetes", N = "n", type.outcome = "rate2", time = "followup")
intervention_viz(summary_diabetes, type.outcome = "rate2")
}

}
