% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma.prior.R
\name{nma.prior}
\alias{nma.prior}
\title{Determine the appropriate value to be used for default priors}
\usage{
nma.prior(data.nma, outcome, scale, N, sd = NULL, time = NULL)
}
\arguments{
\item{data.nma}{A data object created by running data.prep()}

\item{outcome}{A string for the column name containing the outcome}

\item{scale}{A string indicating the scale on which the relative treatment effects are being measures}

\item{N}{A string for the column name containing the sample sizes}

\item{sd}{A string for the column name containing the standard errors (only required for continuous outcomes)}

\item{time}{A string for the column names containing the follow-up times or person-time at risk for rate or rate2 data}
}
\value{
The largest MLE to be used in determining default priors for mu, d and sigma
}
\description{
Finds the largest maximum likelihood estimator from individual trials as in van Valkenhoef, G et al. 2012. This function was originally created in BUGSnet (https://github.com/audrey-b/BUGSnet) without documentation or comments. This version has small changes to avoid warnings during the CRAN check.
}
