% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprebinom2logpost.R
\name{rprebinom2logpost}
\alias{rprebinom2logpost}
\title{Calculate log-posterior in R}
\usage{
rprebinom2logpost(testPars, stanObj)
}
\arguments{
\item{testPars}{A named list of parameters to use}

\item{stanObj}{A StanNetRun object to supply the data}
}
\value{
The value of the log-posterior distribution at the test parameters and data
}
\description{
Computes the log-posterior for a non-centred random effect model with binomial likelihood using the cloglog link and default priors
}
