% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPars.R
\name{simPars}
\alias{simPars}
\title{Simulate parameters for unit testing the log posteriors}
\usage{
simPars(ns, nt, nsim = 10, effects, max.delta)
}
\arguments{
\item{ns}{Tne number of studies}

\item{nt}{The number of unique treatments}

\item{nsim}{The number parameter lists to simulate}

\item{effects}{A character string either "fixed" or "random" indicating what model to simulate parameters for}

\item{max.delta}{The parameter for the standard deviation of the distributions used to simulate parameters mu and d2}
}
\value{
A list of named lists of parameters
}
\description{
Simulate parameters for unit testing the log posteriors
}
