% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortNames.r
\name{sortNames}
\alias{sortNames}
\title{Sort column names by study number, then by arm number}
\usage{
sortNames(string, ns)
}
\arguments{
\item{string}{A character string of the prefix for the desired column names}

\item{ns}{The number of studies
This function is used in convertToBUGSnetRun to ensure the ordering of the data is the same between Stan and JAGS output}
}
\description{
Creates a rearranged vector of column names to be used to reorder columns for compatibility between Stan and JAGS sampling output
}
