% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stan.nma.fit.r
\name{Stan.nma.fit}
\alias{Stan.nma.fit}
\title{Assess NMA model fit}
\usage{
Stan.nma.fit(nma, plot.pD = TRUE, plot.DIC = TRUE, plot.Dres = TRUE, ...)
}
\arguments{
\item{nma}{A BUGSnetRun object produced by running \code{nma.run()} or \code{convertToBUGSnetRun()}}

\item{plot.pD}{Whether to include the effective number of parameters on the plot, default is \code{TRUE}}

\item{plot.DIC}{Whether to include the DIC on the plot, default is \code{TRUE}}

\item{plot.Dres}{Whether to include the posterior mean of the residual deviance on the plot, default is \code{TRUE}}

\item{...}{Graphical arguments such as \verb{main=} to be passed to \code{plot()}}
}
\value{
\code{DIC} - The Deviance Information Criteria
\code{leverage} - A vector of leverages calculated for each data point (one per arm per study)
\code{w} - A vector of each data point's contribution to the posterior mean deviance
\code{pmdev} - A vector of the posterior mean residual deviance for each data point
\code{Dres} - The posterior mean of the residual deviance
\code{pD} - The effective number of parameters, calculated as the sum of \code{leverage}
}
\description{
Computes the Deviance Information Criteria, effective number of parameters, and posterior mean of the residual deviance for a given model, and produces a rainbow plot (called a leverage plot in the NICE Technical Support Document 2). Points outside of the purple curve can be said to contribute to the model's poor fit.
}
