% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanNet-package.R
\name{StanNet}
\alias{StanNet}
\title{StanNet: Bayesian Network Meta-Analysis Using Stan}
\description{
StanNet is an R package to conduct Bayesian network meta-analyses in compliance with best practice and reporting guidelines.
Bayesian analyses are conducted with Stan code generated automatically based on the user's inputs. Outputs are highly customizable and include plots for data visualization and exploration and extensive visualizations for NMA results including rainbow plots, SUCRA plots, rankograms, forest plots, and league tables and heatplots.

To use this package, you must have Stan installed on your computer.
}
